BUG: Fix incorrect output extent of volume reconstructor
Add an optional argument to VolumeReconstructor to change the output spacing slightly so the output extent contains all frames without exceeding the voxel lattice. The output extent can extend beyond existing data when using ceil, or not include some of the data when using floor, which produces holes in the image.